+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file. This projec
 ## 2.2.28 (08/26/2016)
 - HOTFIX: Fixed mil doc blank fields test
 - HOTFIX: Fixed mil email tests
+- HOTFIX: Fixed mil scra tests
 
 ## 2.2.27 (08/25/2016)
 - HOTFIX: Fixed some broken govt tests